% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{pbmc3k-package}
\alias{pbmc3k-package}
\title{pbmc3k: Raw and Processed Matrices of the PBMC 3k Dataset}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Paul Hoffman \email{hoff0792@alumni.umn.edu} (\href{https://orcid.org/0000-0002-7693-8957}{ORCID})

Other contributors:
\itemize{
  \item 10x Genomics, Inc. [copyright holder]
}

}
\keyword{internal}
\keyword{package}
