% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pbmc3k}
\alias{pbmc3k}
\title{PBMC 3k}
\format{
A \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with the raw,
unfiltered PBMC 3k counts
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k
}
\description{
2,700 peripheral blood mononuclear cells (PBMC) from 10X genomics; this is
effectively what one would get with
\code{\link[Seurat:Read10X]{Seurat::Read10X}()}
}
\keyword{datasets}
