% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pbmc3k.sce.logcounts}
\alias{pbmc3k.sce.logcounts}
\title{Log-Normalized PBMC 3k: \pkg{SingleCellExperiment} Edition}
\format{
A \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with the
log-normalized PBMC 3k counts
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k.sce.logcounts
}
\description{
Log-Normalized PBMC 3k: \pkg{SingleCellExperiment} Edition
}
\note{
There is not a raw, filtered matrix following a Bioconductor workflow
as \pkg{scuttle} does not filter the raw counts matrix in the same way that
\code{\link[SeuratObject:CreateSeuratObject]{SeuratObject::CreateSeuratObject}()} does
}
\references{
\href{https://bioconductor.org/books/3.17/OSCA.intro/analysis-overview.html}{Orchestrating Single-Cell Analysis with Bioconductor version 1.8.0}
}
\seealso{
\code{\link[scuttle:perCellQCMetrics]{scuttle::perCellQCMetrics}()},
\code{\link[scuttle:quickPerCellQC]{scuttle::quickPerCellQC}()},
\code{\link[scuttle:perCellQCFilters]{scuttle::perCellQCFilters}()},
\code{\link[scuttle:isOutlier]{scuttle::isOutlier}()}

Log-normalized matrices:
\code{\link{pbmc3k.seurat.norm}}
}
\concept{norm}
\concept{sce}
\keyword{datasets}
