% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pbmc3k.seurat.counts}
\alias{pbmc3k.seurat.counts}
\title{Raw PBMC 3k Counts: \CRANpkg{Seurat} Edition}
\format{
A \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with the raw,
filtered PBMC 3k counts
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k.seurat.counts
}
\description{
Raw PBMC 3k Counts: \CRANpkg{Seurat} Edition
}
\references{
\href{https://satijalab.org/seurat/articles/pbmc3k_tutorial}{Seurat PBMC 3k Guided Clustering Tutorial}
}
\seealso{
\code{\link[SeuratObject:CreateSeuratObject]{SeuratObject::CreateSeuratObject}()},
\code{\link[SeuratObject:CreateAssayObject]{SeuratObject::CreateAssayObject}()}

Seurat-processed matrices:
\code{\link{pbmc3k.seurat.norm}}
}
\concept{raw}
\concept{seurat}
\keyword{datasets}
