% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pbmc3k.seurat.norm}
\alias{pbmc3k.seurat.norm}
\title{Log-Normalized PBMC 3k: \CRANpkg{Seurat} Edition}
\format{
A \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} with the
log-normalized PBMC 3k counts
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc3k.seurat.norm
}
\description{
Log-Normalized PBMC 3k: \CRANpkg{Seurat} Edition
}
\references{
\href{https://satijalab.org/seurat/articles/pbmc3k_tutorial}{Seurat PBMC 3k Guided Clustering Tutorial}
}
\seealso{
\code{\link[Seurat:NormalizeData]{Seurat::NormalizeData}()}

Seurat-processed matrices:
\code{\link{pbmc3k.seurat.counts}}

Log-normalized matrices:
\code{\link{pbmc3k.sce.logcounts}}
}
\concept{norm}
\concept{seurat}
\keyword{datasets}
